lindsey-website/scripts/requirements.txt
psycopg2-binary
python-dotenv
Pillow


requests